+++ src/audio/intelhda.rs
@@ -1,8 +1,10 @@
 use core::ptr::{read, write};
 
 use common::memory::*;
-use common::pci::*;
 use common::scheduler::*;
+
+use drivers::pciconfig::*;
+use drivers::pio::*;
 
 use programs::common::*;
