+++ src/drivers/ps2.rs
@@ -1,12 +1,17 @@
-use common::pio::*;
-
 use drivers::keyboard::keyboard_interrupt;
 use drivers::mouse::mouse_interrupt;
-
-use programs::common::*;
+use drivers::pio::*;
 
 pub struct PS2 {
     status: PIO8
+}
+
+impl PS2 {
+    pub fn new() -> PS2 {
+        return PS2 {
+            status: PIO8::new(0x64)
+        };
+    }
 }
 
 impl SessionItem for PS2 {
