+++ src/drivers/serial.rs
@@ -1,4 +1,4 @@
-use common::pio::*;
+use drivers::pio::*;
 
 use programs::common::*;
