+++ src/network/rtl8139.rs
@@ -1,7 +1,8 @@
 use common::memory::*;
-use common::pci::*;
-use common::pio::*;
 use common::scheduler::*;
+
+use drivers::pciconfig::*;
+use drivers::pio::*;
 
 use network::common::*;
 use network::scheme::*;
