+++ src/usb/uhci.rs
@@ -3,9 +3,10 @@ use core::ptr::{read, write};
 
 use common::context::*;
 use common::memory::*;
-use common::pci::*;
-use common::pio::*;
 use common::scheduler::*;
+
+use drivers::pciconfig::*;
+use drivers::pio::*;
 
 use programs::common::*;
